+++ src/chanjo2/meta/handle_d4.py
@@ -1,6 +1,8 @@
+from decimal import Decimal
 from statistics import mean
 from typing import List, Optional, Tuple
 
+from numpy import ndarray
 from pyd4 import D4File
 from sqlmodel import Session
 
@@ -46,9 +48,36 @@ def intervals_coverage(
     return intervals_cov
 
 
-def genes_coverage(d4_file: D4File, genes: List[SQLGene]) -> List[CoverageInterval]:
-    """Return coverage over a list of genes."""
+def evaluate_region_completeness(
+    d4_file: D4File,
+    chromosome: str,
+    start: int,
+    stop: int,
+    completeness_threholds: Optional[List[int]],
+) -> List[Tuple[int, Decimal]]:
+    """Use NumPy to Calculate the coverage completeness over a list of threshold values for a chromosomal region."""
+
+    if completeness_threholds is None:
+        return []
+
+    region_size: int = stop - start
+    per_base_depth: ndarray = d4_file.load_to_np(f"{chromosome}:{start}-{stop}")
+
+    completeness_values: List[Tuple[int, Decimal]] = []
+
+    for threshold in completeness_threholds:
+        completeness = Decimal((per_base_depth > threshold).sum() / region_size)
+        completeness_values.append((threshold, completeness))
+
+    return completeness_values
+
+
+def genes_coverage_completeness(
+    d4_file: D4File, genes: List[SQLGene], completeness_threholds: List[Optional[int]]
+) -> List[CoverageInterval]:
+    """Return mean coverage and coverage completeness over a list of genes."""
     genes_cov: List[CoverageInterval] = []
+
     for gene in genes:
         genes_cov.append(
             CoverageInterval(
@@ -59,6 +88,13 @@ def genes_coverage(d4_file: D4File, genes: List[SQLGene]) -> List[CoverageInterv
                 start=gene.start,
                 end=gene.stop,
                 mean_coverage=d4_file.mean((gene.chromosome, gene.start, gene.stop)),
+                completeness=evaluate_region_completeness(
+                    d4_file=d4_file,
+                    chromosome=gene.chromosome,
+                    start=gene.start,
+                    stop=gene.stop,
+                    completeness_threholds=completeness_threholds,
+                ),
             )
         )
     return genes_cov