+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Filter genes, transcripts and exons by Ensemble id, HGNC id, HGNC symbol
 - Demo genes, transcripts and exons loaded on demo instance startup
 - Return coverage over a list of genes for a sample in the database
+- Return coverage and coverage completeness (custom thresholds) over a list of genes for a sample in the database
 - Return coverage over a list of genes (transcripts only) for a sample in the database
 
 ### Fixed